chore: move prepare_source fn to lsp from nargo
